Crobat and Gallade Programs Added

+++ SerialPrograms/Source/NintendoSwitch/Commands/NintendoSwitch_Commands_PushButtons.h
@@ -31,15 +31,23 @@ void pbf_mash_button            (const BotBaseContext& context, Button button, u
 void start_program_flash        (const BotBaseContext& context, uint16_t ticks);
 void grip_menu_connect_go_home  (const BotBaseContext& context);
 
+// void pbf_controller_state(
+//     const BotBaseContext& context,
+//     Button button,
+//     DpadPosition position,
+//     uint8_t left_x, uint8_t left_y,
+//     uint8_t right_x, uint8_t right_y,
+//     uint8_t ticks
+// );
+
 void pbf_controller_state(
     const BotBaseContext& context,
     Button button,
     DpadPosition position,
     uint8_t left_x, uint8_t left_y,
     uint8_t right_x, uint8_t right_y,
-    uint8_t ticks
+    uint16_t ticks
 );
-
